requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
Django==2.2.4
django-cors-headers==3.1.0
django-extensions==2.2.1
django-filter==2.2.0
django-model-utils==3.2.0
django-templated-mail==1.1.1
djangorestframework==3.10.2
djoser==2.0.3
Markdown==3.1.1
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
uWSGI==2.0.18
